refactor(webapp): add EndpointsChips component
